# docs/swagger.yaml
basePath: /
definitions:
...
  entity.Menu:
    properties:
      children:
        items:
          $ref: '#/definitions/entity.Menu'
        type: array
      display:
        maxLength: 100
        minLength: 2
        type: string
      icon:
        maxLength: 50
        type: string
      id:
        type: string
      name:
        maxLength: 50
        minLength: 2
        type: string
      orderNum:
        minimum: 0
        type: integer
      parentId:
        type: string
      permission:
        maxLength: 100
        type: string
      status:
        enum:
        - "0"
        - "1"
        type: string
      type:
        description: M目录 C菜单 F按钮
        enum:
        - M
        - C
        - F
        type: string
      url:
        maxLength: 255
        type: string
      visible:
        enum:
        - "0"
        - "1"
        type: string
    required:
    - display
    - name
    - type
    type: object
  entity.Notice:
    properties:
      content:
        type: string
      createdBy:
        type: string
      createdDate:
        type: string
      endTime:
        type: string
      id:
        type: string
      startTime:
        type: string
      title:
        type: string
      updatedBy:
        type: string
      updatedDate:
        type: string
      userDomain:
        type: string
    type: object
  entity.Post:
    properties:
      code:
        type: string
      createdBy:
        type: string
      createdDate:
        type: string
      id:
        type: string
      name:
        type: string
      remark:
        type: string
      sort:
        type: integer
      status:
        type: string
      updatedBy:
        type: string
      updatedDate:
        type: string
    type: object
  entity.Role:
    properties:
      createdDate:
        type: string
      display:
        maxLength: 100
        minLength: 2
        type: string
      id:
        type: string
      isSystem:
        type: boolean
      name:
        maxLength: 50
        minLength: 2
        type: string
      status:
        enum:
        - active
        - inactive
        type: string
    required:
    - display
    - name
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      restCode:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: simonchen453@gmail.com
    name: API Support
  description: 基于 Clean Architecture 的企业级后台管理系统 API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Pro Golang API
  version: "1.0"
paths:
...
  /api/v1/auth/userinfo:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取当前用户信息
      tags:
      - user
...
  /api/v1/monitor/online/{id}:
    delete:
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 强退用户
      tags:
      - monitor
...
  /api/v1/system/menu:
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/entity.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增菜单
      tags:
      - system/menu
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜单信息
        in: body
        name: menu
        required: true
        schema:
          $ref: '#/definitions/entity.Menu'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新菜单
      tags:
      - system/menu
  /api/v1/system/menu/{id}:
    delete:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除菜单
      tags:
      - system/menu
    get:
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取菜单详情
      tags:
      - system/menu
...
  /api/v1/system/notice:
    post:
      consumes:
      - application/json
      parameters:
      - description: 公告信息
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/entity.Notice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增通知公告
      tags:
      - system/notice
    put:
      consumes:
      - application/json
      parameters:
      - description: 公告信息
        in: body
        name: notice
        required: true
        schema:
          $ref: '#/definitions/entity.Notice'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新通知公告
      tags:
      - system/notice
  /api/v1/system/notice/{id}:
    delete:
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除通知公告
      tags:
      - system/notice
    get:
      parameters:
      - description: 公告ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取通知公告详情
      tags:
      - system/notice
...
  /api/v1/system/post:
    post:
      consumes:
      - application/json
      parameters:
      - description: 岗位信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增岗位
      tags:
      - system/post
    put:
      consumes:
      - application/json
      parameters:
      - description: 岗位信息
        in: body
        name: post
        required: true
        schema:
          $ref: '#/definitions/entity.Post'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新岗位
      tags:
      - system/post
  /api/v1/system/post/{id}:
    delete:
      parameters:
      - description: 岗位ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除岗位
      tags:
      - system/post
    get:
      parameters:
      - description: 岗位ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取岗位详情
      tags:
      - system/post
...
  /api/v1/system/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/entity.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 新增角色
      tags:
      - system/role
    put:
      consumes:
      - application/json
      parameters:
      - description: 角色信息
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/entity.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新角色
      tags:
      - system/role
  /api/v1/system/role/{id}:
    delete:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除角色
      tags:
      - system/role
    get:
      parameters:
      - description: 角色ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取角色详情
      tags:
      - system/role
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"